Cell.java : Factory Method Pattern
As Factory Method Pattern applied(CellCore - Abstract, CellCoreOriginal
- Concrete), Client class(Cell) create cellcore with createCellCore()
  method.
